FEAT

+++ src/main/java/com/jeontongju/seller/dto/temp/FeignFormat.java
@@ -13,13 +13,13 @@ import lombok.Data;
 @Data
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class FeignFormat<T> {
-    private final Long code;
+    private final Integer code;
     private final String message;
     private final String detail;
     private final String failure;
     private final T data;
     @Builder
-    public FeignFormat(Long code, String message, String detail, String failure, Object data) {
+    public FeignFormat(Integer code, String message, String detail, String failure, Object data) {
         this.code = code;
         this.message = message;
         this.detail = detail;
